Fixed archiving
