+++ src/main/java/de/muenchen/mobidam/s3/S3RouteBuilder.java
@@ -93,7 +93,7 @@ public class S3RouteBuilder extends RouteBuilder {
                 .toD(String.format(
                         "aws2-s3://${header.%1$s}?accessKey=RAW(${header.%2$s})&secretKey=RAW(${header.%3$s})&region={{camel.component.aws2-s3.region}}&operation=${header.%4$s}&overrideEndpoint=true&uriEndpointOverride={{camel.component.aws2-s3.override-endpoint}}&${header.%5$s}",
                         CommonConstants.HEADER_BUCKET_NAME, CommonConstants.HEADER_ACCESS_KEY, CommonConstants.HEADER_SECRET_KEY, AWS2S3Constants.S3_OPERATION,
-                        Constants.PARAMETER_ARCHIVED))
+                        Constants.PARAMETER_PREFIX))
                 .process("archiveOperationWrapper")
                 .toD(String.format(
                         "aws2-s3://${header.%1$s}?accessKey=RAW(${header.%2$s})&secretKey=RAW(${header.%3$s})&region={{camel.component.aws2-s3.region}}&operation=%4$s&overrideEndpoint=true&uriEndpointOverride={{camel.component.aws2-s3.override-endpoint}}",
